+++ modules/gui/macosx/library/video-library/VLCLibraryVideoDataSource.m
@@ -63,7 +63,7 @@ NSString * const VLCLibraryVideoDataSourceDisplayedCollectionChangedNotification
     return self;
 }
 
-- (NSUInteger)indexOfMediaItem:(const NSUInteger)libraryId inArray:(NSArray const *)array
+- (NSUInteger)indexOfMediaItem:(int64_t)libraryId inArray:(NSArray const *)array
 {
     return [array indexOfObjectPassingTest:^BOOL(VLCMediaLibraryMediaItem * const findMediaItem, const NSUInteger __unused idx, BOOL * const __unused stop) {
         NSAssert(findMediaItem != nil, @"Collection should not contain nil media items");
